move src/parsers.h to src/parsers/parsers.h
